requirements.txt
Flask==3.1.2
google-generativeai==0.8.5
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.1
itsdangerous==2.2.0
blinker==1.9.0
gunicorn==22.0.0
